flask-app/requirements.txt
Flask==2.3.6
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-exporter-otlp==1.17.0
